FIX: Checking if createdVersion is null in case job fails before creating version

+++ src/main/java/edu/nd/crc/safa/features/jobs/entities/app/CommitJob.java
@@ -85,7 +85,7 @@ public abstract class CommitJob extends AbstractJob {
      */
     @Override
     protected void jobFailed(Exception error) throws RuntimeException, IOException {
-        if (this.deleteProjectOnFail) {
+        if (this.deleteProjectOnFail && createdProjectVersion != null) {
             this.getDbLogger().log("Job failed, deleting job.");
             Project project = createdProjectVersion.getProject();
             getServiceProvider().getProjectService().deleteProject(project);
